fix(DP-1060): fold demographics into query preview text everywhere
Recent searches, the query history table and the results header rendered a
saved query via the concept-only rule renderer, which ignores the
demographics block. A demographics-only query (age/sex/location, no concept
rules) therefore rendered as an empty string.

queryToText is now the single entry point for turning a definition into
preview text: it renders the concept rules then folds the demographic subject
in by replacing the leading "People" noun, so the subject scopes the whole
"who ..." clause regardless of OR/AND/mixed combinators (matching the AND
intersection of demographics with the rule tree). The concept-only renderer is
queryRulesToText.

The editable search input must stay concept-only (its text round-trips to
concept matching), so the store computes queryAsText with
queryToText(def, { includeDemographics: false }). Every display site
(SearchOverlay, QueriesTable, QueryResultsPage, CohortQueryPreview) uses the
default demographics-inclusive rendering, independent of feature flags.

The demographic-subject helpers live in queryBuilder.ts alongside the rule
renderer (no separate demographicsText util), and formatSexNoun guards against
legacy definitions with a missing sex array.
